[IFC][Integration] Introduce InlineContentPainter https://bugs.webkit.org/show_bug.cgi?id=244339
Reviewed by Antti Koivisto.

Move painting logic from LineLayout to InlineContentPainter.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/layout/integration/inline/LayoutIntegrationInlineContentPainter.cpp: Added.
(WebCore::LayoutIntegration::InlineContentPainter::InlineContentPainter):
(WebCore::LayoutIntegration::LayerPaintScope::LayerPaintScope):
(WebCore::LayoutIntegration::LayerPaintScope::includes):
(WebCore::LayoutIntegration::InlineContentPainter::paint):
(WebCore::LayoutIntegration::InlineContentPainter::flippedContentOffsetIfNeeded const):
* Source/WebCore/layout/integration/inline/LayoutIntegrationInlineContentPainter.h: Added.
* Source/WebCore/layout/integration/inline/LayoutIntegrationLineLayout.cpp:
(WebCore::LayoutIntegration::LineLayout::paint):
(WebCore::LayoutIntegration::LayerPaintScope::LayerPaintScope): Deleted.
(WebCore::LayoutIntegration::LayerPaintScope::includes): Deleted.
(): Deleted.
* Source/WebCore/layout/integration/inline/LayoutIntegrationLineLayout.h:

Canonical link: https://commits.webkit.org/253801@main
